Add files via upload
I added to the update function, to remove the dots from the screen once pacman is in the same location as a dot(eats one).
This caused difficulty at first as I was passing by reference self.dots = self.free_cells. I fixed this by appending all the locations to self.dots in the game file. This mean't that I was passing by reference anymore and I was not removing from the self.free_cells list and just the self.dots list.